[REFACTOR] views 파일 분리
