+++ jgw_api/views/__init__.py
@@ -1,0 +1,6 @@
+from .view_borad import *
+from .view_post import *
+from .view_comment import *
+from .view_image import *
+from .view_survey import *
+from .view_check import ping_pong
